feat(FileSharedActions): Remove print callback from BasePreviewSliderFragment
