+++ app/src/main/java/com/infomaniak/drive/ui/fileShared/FileSharedPreviewSliderFragment.kt
@@ -135,6 +135,7 @@ class FileSharedPreviewSliderFragment : BasePreviewSliderFragment(), FileInfoAct
     }
 
     override fun printClicked() {
+        super<BasePreviewSliderFragment>.printClicked()
         previewSliderViewModel.executeDownloadAction(
             activityContext = requireContext(),
             downloadAction = DownloadAction.PRINT_PDF,
